extract and reuse HandleInvalid across all models that require it
