+++ mleap-core/src/main/scala/ml/combust/mleap/core/feature/StringIndexerModel.scala
@@ -3,33 +3,6 @@ package ml.combust.mleap.core.feature
 import ml.combust.mleap.core.Model
 import ml.combust.mleap.core.types.{ScalarType, StructType}
 
-sealed trait HandleInvalid {
-  def asParamString: String
-}
-
-object HandleInvalid {
-  val default = Error
-
-  case object Error extends HandleInvalid {
-    override def asParamString: String = "error"
-  }
-
-  case object Skip extends HandleInvalid {
-    override def asParamString: String = "skip"
-  }
-
-  case object Keep extends HandleInvalid {
-    override def asParamString: String = "keep"
-  }
-
-  def fromString(value: String): HandleInvalid = value match {
-    case "error" => HandleInvalid.Error
-    case "skip" => HandleInvalid.Skip
-    case "keep" => HandleInvalid.Keep
-    case _ => throw new IllegalArgumentException(s"Invalid handler: $value")
-  }
-}
-
 /** Class for string indexer model.
   *
   * String indexer converts a string into an integer representation.
